move set_common_showutils to targets/utils
