+++ siliconcompiler/targets/asap7_demo.py
@@ -1,4 +1,5 @@
 import siliconcompiler
+from . import utils
 
 def make_docs():
     '''
@@ -24,7 +25,7 @@ def setup(chip):
     chip.use(asap7sc7p5t)
 
     #2. Setup default show tools
-    siliconcompiler.utils.set_common_showtools(chip)
+    utils.set_common_showtools(chip)
 
     #3. Select default flow/PDK
     chip.set('option', 'mode', 'asic')
