+++ siliconcompiler/targets/skywater130_demo.py
@@ -1,4 +1,5 @@
 import siliconcompiler
+from . import utils
 
 ############################################################################
 # DOCS
@@ -37,7 +38,7 @@ def setup(chip):
     chip.use(oh_tapeout)
 
     #2. Setup default show tools
-    siliconcompiler.utils.set_common_showtools(chip)
+    utils.set_common_showtools(chip)
 
     #3. Set default targets
     chip.set('option', 'mode', 'asic')
